+++ src/main/java/org/bytedeco/javacpp/annotation/ByVal.java
@@ -19,4 +19,7 @@ import org.bytedeco.javacpp.tools.Generator;
  */
 @Documented @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD, ElementType.PARAMETER})
-public @interface ByVal { }
+public @interface ByVal {
+    /** A C++ code snippet to use when null is passed. By default {@link NullPointerException} is thrown. */
+    String nullValue() default "";
+}