+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Introduce a `nullValue` to `@ByRef` and `@ByVal` annotations to let us specify what to do when passed `null`
  * Add properties for `android-arm64`, `android-x86_64`, and `linux-arm64` platforms
  * Add slow but generic `Indexer.putDouble()` to complement existing `Indexer.getDouble()` (useful for testing)
  * Fix and enhance in various ways the support of `Parser` and `Generator` for function pointers, virtual functions, and abstract classes